// lib/data/airports.generated.ts
export interface RawAirport {
  iata: string;
  city: string;
  name: string;
  country: string;
  lat: number;
  lon: number;
  aliases?: string[];
}